Tests/LibWeb/Text/expected/window-scrollTo.txt

  The page has been scrolled to y: 1616
